requirements.txt
flask==2.2.5
openai-whisper
librosa
numpy
scikit-learn
soundfile
ffmpeg-python
torch
